Split Name::Create and Name::Lifeform tests out of name_test.rb (#4708)
Moves test_find_or_create_name_and_parents into
test/models/name/create_test.rb (matching app/models/name/create.rb)
and test_lichen into test/models/name/lifeform_test.rb (matching
app/models/name/lifeform.rb).

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
