+++ test/models/name/create_test.rb
@@ -1,0 +1,63 @@
+# frozen_string_literal: true
+
+require("test_helper")
+
+# Tests for Name::Create (app/models/name/create.rb)
+class Name::CreateTest < UnitTestCase
+  def test_find_or_create_name_and_parents
+    # Coprinus comatus already has an author.
+    # Create new subspecies Coprinus comatus v. bogus and make sure it doesn't
+    # create a duplicate species if one already exists.
+    # Saw this bug 20080114 -JPH
+    result = Name.find_or_create_name_and_parents(
+      rolf, "Coprinus comatus v. bogus (With) Author"
+    )
+    assert_equal(3, result.length)
+    assert_equal(names(:coprinus).id, result[0].id)
+    assert_equal(names(:coprinus_comatus).id, result[1].id)
+    assert_nil(result[2].id)
+    assert_equal("Coprinus", result[0].text_name)
+    assert_equal("Coprinus comatus", result[1].text_name)
+    assert_equal("Coprinus comatus var. bogus", result[2].text_name)
+    assert_equal(names(:coprinus).author, result[0].author)
+    assert_equal("(O.F. Müll.) Pers.", result[1].author)
+    assert_equal("(With) Author", result[2].author)
+
+    # Conocybe filaris does not have an author.
+    result = Name.find_or_create_name_and_parents(
+      rolf, "Conocybe filaris var bogus (With) Author"
+    )
+    assert_equal(3, result.length)
+    assert_equal(names(:conocybe).id, result[0].id)
+    assert_equal(names(:conocybe_filaris).id, result[1].id)
+    assert_nil(result[2].id)
+    assert_equal("Conocybe", result[0].text_name)
+    assert_equal("Conocybe filaris", result[1].text_name)
+    assert_equal("Conocybe filaris var. bogus", result[2].text_name)
+    assert_equal("", result[0].author)
+    assert_equal("", result[1].author)
+    assert_equal("(With) Author", result[2].author)
+
+    # Agaricus fixture does not have an author.
+    result = Name.find_or_create_name_and_parents(rolf, "Agaricus L.")
+    assert_equal(1, result.length)
+    assert_equal(names(:agaricus).id, result[0].id)
+    assert_equal("Agaricus", result[0].text_name)
+    assert_equal("L.", result[0].author)
+
+    # Agaricus does not have an author.
+    result = Name.find_or_create_name_and_parents(
+      rolf, "Agaricus abra f. cadabra (With) Another Author"
+    )
+    assert_equal(3, result.length)
+    assert_equal(names(:agaricus).id, result[0].id)
+    assert_nil(result[1].id)
+    assert_nil(result[2].id)
+    assert_equal("Agaricus", result[0].text_name)
+    assert_equal("Agaricus abra", result[1].text_name)
+    assert_equal("Agaricus abra f. cadabra", result[2].text_name)
+    assert_equal("", result[0].author)
+    assert_equal("", result[1].author)
+    assert_equal("(With) Another Author", result[2].author)
+  end
+end
